Move study overview type to home controller types
