+++ controllers/home/types.ts
@@ -1,4 +1,12 @@
-import { StudyOverview } from "@/types/study";
+import { Tag } from "@/types/study";
+
+export type StudyOverview = {
+  id: number;
+  title: string;
+  description: string;
+  remainTime: string;
+  tags: Tag[];
+};
 
 export type HomePage = {
   data: {
